Routine.Samples.SmartObject is converted to .net 5
