--- samples/Routine.Samples.SmartObject/Global.asax
@@ -1,1 +1,0 @@
-﻿<%@ Application Codebehind="Global.asax.cs" Inherits="Routine.Samples.SmartObject.Global" Language="C#" %>
