--- samples/Routine.Samples.SmartObject/Web.config
@@ -1,13 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-
-<!--
-  For more information on how to configure your ASP.NET application, please visit
-  https://go.microsoft.com/fwlink/?LinkId=169433
-  -->
-<configuration>
-  <system.web>
-    <compilation debug="true" targetFramework="4.8"/>
-    <httpRuntime targetFramework="4.8"/>
-  </system.web>
-
-</configuration>
